+++ common/src/main/java/net/satisfy/bakery/core/block/TrayBlock.java
@@ -1,46 +1,27 @@
 package net.satisfy.bakery.core.block;
 
-import net.minecraft.ChatFormatting;
-import net.minecraft.Util;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.HorizontalDirectionalBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import net.satisfy.bakery.core.registry.StorageTypeRegistry;
-import net.satisfy.farm_and_charm.core.util.GeneralUtil;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.function.Supplier;
-
 public class TrayBlock extends CakeStandBlock {
+    private static final VoxelShape SHAPE = Shapes.box(0.1875, 0.0, 0.125, 0.8125, 0.3125, 0.875);
+
     public TrayBlock(BlockBehaviour.Properties settings) {
         super(settings);
     }
 
-    private static final Supplier<VoxelShape> voxelShapeSupplier = () -> Shapes.or(Shapes.box(0.25, 0.0, 0.1875, 0.75, 0.0625, 0.8125), Shapes.box(0.1875, 0.0625, 0.125, 0.8125, 0.3125, 0.875));
-
-    public static final Map<Direction, VoxelShape> SHAPE = Util.make(new HashMap<>(), map -> {
-        for (Direction direction : Direction.Plane.HORIZONTAL.stream().toList()) {
-            map.put(direction, GeneralUtil.rotateShape(Direction.NORTH, direction, voxelShapeSupplier.get()));
-        }
-    });
-
     @Override
     public @NotNull VoxelShape getShape(BlockState state, BlockGetter world, BlockPos pos, CollisionContext context) {
-        return SHAPE.get(state.getValue(HorizontalDirectionalBlock.FACING));
+        return SHAPE;
     }
 
     @Override
@@ -60,16 +41,8 @@ public class TrayBlock extends CakeStandBlock {
 
     @Override
     public int getSection(Float f, Float y) {
-        int nSection;
-        float oneS = 1.0f / 9;
-
-        nSection = (int) (f / oneS);
-
-        return 8 - nSection;
-    }
-
-    @Override
-    public void appendHoverText(ItemStack itemStack, Item.TooltipContext tooltipContext, List<Component> list, TooltipFlag tooltipFlag) {
-        list.add(Component.translatable("tooltip.farm_and_charm.canbeplaced").withStyle(ChatFormatting.GRAY));
+        float oneSection = 1.0f / 9;
+        int sectionIndex = (int) (f / oneSection);
+        return 8 - sectionIndex;
     }
 }